+++ modules/device/src/device.cpp
@@ -18,6 +18,7 @@
 #include <ifm3d/common/logging/log.h>
 #include <ifm3d/device/device.h>
 #include <ifm3d/device/ifm_network_device.h>
+#include <ifm3d/device/o3c.h>
 #include <ifm3d/device/o3d.h>
 #include <ifm3d/device/o3r.h>
 #include <ifm3d/device/o3x.h>
@@ -112,6 +113,8 @@ const int ifm3d::DEV_O3D_MIN = 1;
 const int ifm3d::DEV_O3D_MAX = 255;
 const int ifm3d::DEV_O3R_MIN = 768;
 const int ifm3d::DEV_O3R_MAX = 1023;
+const int ifm3d::DEV_O3C_MIN = 1280;
+const int ifm3d::DEV_O3C_MAX = 1535;
 const int ifm3d::DEV_O3X_MIN = 512;
 const int ifm3d::DEV_O3X_MAX = 767;
 const std::string ifm3d::ASSUME_DEVICE =
@@ -139,7 +142,9 @@ const unsigned int ifm3d::O3X_DISTANCE_NOISE_IMAGE_SUPPORT_MAJOR = 1;
 const unsigned int ifm3d::O3X_DISTANCE_NOISE_IMAGE_SUPPORT_MINOR = 1;
 const unsigned int ifm3d::O3X_DISTANCE_NOISE_IMAGE_SUPPORT_PATCH = 190;
 
-const ifm3d::SemVer O3R_MINIMUM_FIRWARE_SUPPORTED(0, 13, 13);
+const ifm3d::SemVer O3R_MINIMUM_FIRMWARE_SUPPORTED(0, 13, 13);
+// TODO : Update once O3C minimum firmware is defined
+const ifm3d::SemVer O3C_MINIMUM_FIRMWARE_SUPPORTED(0, 13, 13);
 
 //================================================
 // Function for Searching Devices on Network
@@ -183,9 +188,9 @@ ifm3d::Device::MakeShared(const std::string& ip,
       if (base->AmI(DeviceFamily::O3R))
         {
           if (base->CheckMinimumFirmwareVersion(
-                O3R_MINIMUM_FIRWARE_SUPPORTED.major_num,
-                O3R_MINIMUM_FIRWARE_SUPPORTED.minor_num,
-                O3R_MINIMUM_FIRWARE_SUPPORTED.patch_num))
+                O3R_MINIMUM_FIRMWARE_SUPPORTED.major_num,
+                O3R_MINIMUM_FIRMWARE_SUPPORTED.minor_num,
+                O3R_MINIMUM_FIRMWARE_SUPPORTED.patch_num))
             {
               LOG_VERBOSE("Instantiating O3R...");
               return std::make_shared<ifm3d::O3R>(ip, xmlrpc_port);
@@ -194,7 +199,26 @@ ifm3d::Device::MakeShared(const std::string& ip,
           const std::string error_msg =
             fmt::format("Please update the firmware, minimum firmware "
                         "version required is {}",
-                        O3R_MINIMUM_FIRWARE_SUPPORTED);
+                        O3R_MINIMUM_FIRMWARE_SUPPORTED);
+
+          LOG_VERBOSE(error_msg);
+          throw Error(IFM3D_INVALID_FIRMWARE_VERSION, error_msg);
+        }
+      if (base->AmI(DeviceFamily::O3C))
+        {
+          if (base->CheckMinimumFirmwareVersion(
+                O3C_MINIMUM_FIRMWARE_SUPPORTED.major_num,
+                O3C_MINIMUM_FIRMWARE_SUPPORTED.minor_num,
+                O3C_MINIMUM_FIRMWARE_SUPPORTED.patch_num))
+            {
+              LOG_VERBOSE("Instantiating O3C...");
+              return std::make_shared<ifm3d::O3C>(ip, xmlrpc_port);
+            }
+
+          const std::string error_msg =
+            fmt::format("Please update the firmware, minimum firmware "
+                        "version required is {}",
+                        O3C_MINIMUM_FIRMWARE_SUPPORTED);
 
           LOG_VERBOSE(error_msg);
           throw Error(IFM3D_INVALID_FIRMWARE_VERSION, error_msg);
@@ -332,6 +356,10 @@ ifm3d::Device::WhoAmI()
   if (check_device_id(device_id(), ifm3d::DEV_O3R_MIN, ifm3d::DEV_O3R_MAX))
     {
       return DeviceFamily::O3R;
+    }
+  if (check_device_id(device_id(), ifm3d::DEV_O3C_MIN, ifm3d::DEV_O3C_MAX))
+    {
+      return DeviceFamily::O3C;
     }
 
   return DeviceFamily::UNKNOWN;