+++ ChangeLog.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Add `IMUInfoV1` deserializer for O3R IMU data.
+- Add `o3c` device support to ifm3d
 
 ### Fixed
 - Fix `IMUInfoV1.imu_samples` Python binding to return a valid list instead of conversion error.